examples/shop/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
httpx>=0.27
python-dotenv>=1.0
jinja2>=3.1
